line number for files with no newline
The parser sets the line number used for error messages once reaching a
newline.  This fixes the logic for the case that the input file does not
have a newline.

This is only visible on .i files, as the preprocessor adds a newline when
there isn't one on .c files.